Classifier free guidance unconditioned value (#8562)
Fixes #8560.

### Description

This PR adds an argument cfg_fill_value so that users can control the
value that replaces the conditioning tensor during classifier-free
guidance inference.
Previously, this was set to -1, which might not be ideal for certain
application where the conditioning could have -1 as a normal value.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [x] New tests added to cover the changes **modified existing ones
- [x] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [x] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [x] In-line docstrings updated.

---------

Co-authored-by: Virginia Fernandez <virginia.fernandez@kcl.ac.uk>